Immersa table refactoring (#19)
* synced immersatable improvements with old-dev branch
* added missing dependencies
* handled json parsing issue
* added custom control for search bar and column reordering
* handled CSS issues
