+++ superset-frontend/packages/superset-plugin-chart-immersatable/package.json
@@ -50,8 +50,10 @@
     "@visx/vendor": "3.5.0",
     "moment": "^2.25.3",
     "memoize-one": "^5.1.1",
+    "@react-icons/all-files": "^4.1.0",
     "@types/d3-time-format": "^2.1.1",
-    "d3-time-format": "^3.0.0"
+    "d3-time-format": "^3.0.0",
+    "regenerator-runtime": "^0.13.7"
   },
   "peerDependencies": {
     "@superset-ui/chart-controls": "*",
