+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/Styles.ts
@@ -1,0 +1,49 @@
+/* eslint-disable theme-colors/no-literal-colors */
+import { styled } from '@superset-ui/core';
+import { SupersetPluginChartImmersatableStylesProps } from './types';
+
+export const Styles = styled.div<SupersetPluginChartImmersatableStylesProps>`
+  padding: ${({ theme }) => theme.gridUnit * 2}px;
+  border-radius: ${({ theme }) => theme.gridUnit * 2}px;
+  height: ${({ height }) => height}px;
+  width: ${({ width }) => width}px;
+  overflow: auto;
+  h3 {
+    /* You can use your props to control CSS! */
+    margin-top: 0;
+    margin-bottom: ${({ theme }) => theme.gridUnit * 3}px;
+    font-size: ${({ theme, headerFontSize }) =>
+      theme.typography.sizes[headerFontSize]}px;
+    font-weight: ${({ theme, boldText }) =>
+      theme.typography.weights[boldText ? 'bold' : 'normal']};
+  }
+
+  pre {
+    height: ${({ theme, headerFontSize, height }) =>
+      height - theme.gridUnit * 12 - theme.typography.sizes[headerFontSize]}px;
+  }
+`;
+
+const baseStyle = styled.div`
+  border: 1px solid #d1d5db;
+`;
+
+export const ContainerStyled = styled(baseStyle)`
+  border-radius: 1rem;
+  margin: 10px;
+  width: fit-content;
+  overflow: hidden;
+  &:hover {
+    box-shadow: 0 20px 25px -5px rgb(0 0 0 / 0.1),
+      0 8px 10px -6px rgb(0 0 0 / 0.1);
+  }
+`;
+export const HeaderStyled = styled(baseStyle)`
+  padding: 17px 24px;
+  border-top-left-radius: 1rem;
+  border-top-right-radius: 1rem;
+  background: #f3f4f6;
+  font-size: 1.4rem;
+  font-weight: bold !important;
+  color: rgb(107, 114, 128);
+`;
