+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/superset-core-utils/format-column-value.ts
@@ -1,0 +1,53 @@
+import {
+  DataRecordValue,
+  GenericDataType,
+  getNumberFormatter,
+  isProbablyHTML,
+  sanitizeHtml,
+} from '@superset-ui/core';
+import { DataColumnMeta } from '../types';
+import DateWithFormatter from './date-with-formatter';
+
+function formatValue(
+  formatter: DataColumnMeta['formatter'],
+  value: DataRecordValue,
+): [boolean, string] {
+  // render undefined as empty string
+  if (value === undefined) {
+    return [false, ''];
+  }
+  // render null as `N/A`
+  if (
+    value === null ||
+    // null values in temporal columns are wrapped in a Date object, so make sure we
+    // handle them here too
+    (value instanceof DateWithFormatter && value.input === null)
+  ) {
+    return [false, 'N/A'];
+  }
+  if (formatter) {
+    return [false, formatter(value as number)];
+  }
+  if (typeof value === 'string') {
+    return isProbablyHTML(value) ? [true, sanitizeHtml(value)] : [false, value];
+  }
+  return [false, value.toString()];
+}
+
+export function formatColumnValue(
+  column: DataColumnMeta,
+  value: DataRecordValue,
+) {
+  const { dataType, formatter, config = {} } = column;
+  const isNumber = dataType === GenericDataType.NUMERIC;
+  const smallNumberFormatter =
+    config.d3SmallNumberFormat === undefined
+      ? formatter
+      : getNumberFormatter(config.d3SmallNumberFormat);
+  return formatValue(
+    isNumber && typeof value === 'number' && Math.abs(value) < 1
+      ? smallNumberFormatter
+      : formatter,
+    value,
+  );
+}
