+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/superset-core-utils/get-shared-style.ts
@@ -1,0 +1,14 @@
+import { CSSProperties } from 'react';
+import { DataColumnMeta } from '../types';
+
+export const getSharedStyle = (column: DataColumnMeta): CSSProperties => {
+  const { isNumeric, config = {} } = column;
+  const textAlign = config.horizontalAlign
+    ? config.horizontalAlign
+    : isNumeric
+    ? 'right'
+    : 'left';
+  return {
+    textAlign,
+  };
+};
