+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/utils/check-chart-data.ts
@@ -1,0 +1,10 @@
+export const checkChartData = (cellValue: string) => {
+  if (
+    cellValue.startsWith('[[') &&
+    cellValue.endsWith(']]') &&
+    Array.isArray(JSON.parse(cellValue as string))
+  ) {
+    return true;
+  }
+  return false;
+};
