+++ superset-frontend/packages/superset-plugin-chart-immersatable/types/external.d.ts
@@ -17,7 +17,6 @@
  * under the License.
  */
 
-declare module '*.png' {
-  const value: any;
-  export default value;
-}
+declare module '*.png';
+declare module '*.jpg';
+declare module 'regenerator-runtime/runtime';
